main_server/requirements.txt
uvicorn
fastapi
opencv-python
PyWavelets
pillow
scikit-learn
python-multipart